requirements_production.txt
# Core Flask and Extensions - Production Optimized
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-WTF==1.2.1
Flask-Babel==4.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Caching==2.1.0

# Database - PostgreSQL for production
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Forms and Validation
WTForms==3.1.1
email-validator==2.1.0

# Security
Werkzeug==3.0.1
cryptography>=42.0.0
bcrypt>=4.0.0

# HTTP and API
requests==2.31.0
urllib3==2.1.0

# Data Processing
pandas==2.1.4
openpyxl==3.1.2
xlsxwriter==3.1.9

# PDF Generation
reportlab==4.0.7

# Image Processing (Essential only)
Pillow==10.1.0

# Two-Factor Authentication
pyotp==2.8.0
qrcode[pil]==7.4.2

# Environment
python-dotenv==1.0.0

# Production Server
gunicorn==21.2.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
click==8.1.7
itsdangerous==2.1.2
MarkupSafe==2.1.3
Jinja2==3.1.2

# File handling
mimetypes-extended==1.0.0